fix(121): split inject state to separate file
