+++ model/inject_state.go
@@ -1,0 +1,29 @@
+// Copyright 2022 The Serverless Workflow Specification Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package model
+
+// InjectState used to inject static data into state data input.
+type InjectState struct {
+	BaseState
+	// JSON object which can be set as states data input and can be manipulated via filters
+	Data map[string]interface{} `json:"data" validate:"required,min=1"`
+	// State specific timeouts
+	Timeouts *InjectStateTimeout `json:"timeouts,omitempty"`
+}
+
+// InjectStateTimeout defines timeout settings for inject state
+type InjectStateTimeout struct {
+	StateExecTimeout *StateExecTimeout `json:"stateExecTimeout,omitempty"`
+}
